Create README.md
This repository is about web scraper 
It often happens that we want to acquire data, but it is not available as an excel file. The answer to this is web scraper using beautiful soap library.  
In this repository I have shown how beautiful soap can be used to extract data from a website using the example of Wikipiedia